requirements.txt
ascendebug==0.1.0
auto_deploy_utils==0.1.0
auto_tune==0.1.0
blinker==1.8.2
click==8.1.8
dataflow==0.0.1
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
hccl==0.1.0
hccl_parser==0.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
llm_datadist==0.0.1
MarkupSafe==2.1.5
msobjdump==0.1.0
netifaces==0.11.0
op_compile_tool==0.1.0
op_gen==0.1
op_test_frame==0.1
opc_tool==0.1.0
psutil==5.9.5
scapy==2.6.1
schedule_search==0.0.1
show_kernel_debug_data==0.1.0
te==0.4.0
Werkzeug==3.0.6
zipp==3.20.2